+++ package.json
@@ -31,6 +31,7 @@
     "react-loading": "^2.0.3",
     "react-loading-skeleton": "^3.1.0",
     "react-responsive-carousel": "^3.2.23",
+    "react-router-dom": "^6.11.0",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.6",
     "typescript": "^4.9.4",
