+++ src/index.tsx
@@ -1,14 +1,17 @@
 import ReactDOM from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
 import './index.css';
+import { BrowserRouter } from 'react-router-dom';
 import axios from 'axios';
-import App from './App';
+import App from 'App';
 
 axios.defaults.baseURL = process.env.REACT_APP_API_URL;
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
-  <RecoilRoot>
-    <App />
-  </RecoilRoot>
+  <BrowserRouter>
+    <RecoilRoot>
+      <App />
+    </RecoilRoot>
+  </BrowserRouter>
 );
